Initial commit
Shoutout to krabbenkutter for starting this project!

All of the changes I have made are stored in the header of the .ino file. This may change in the future now that the code is on Git. In summary:
-Blink procedure to confirm power on
-Added a curve mode, called transition curve, with mode selection at startup using a button
-Moved the debug code into a function for ease of use
-Excel spreadsheet with curve mode values and charts
-Description with instructions, method of operation, notes, and changelog.

This should all work in theory, I have yet to test it, but will be able to put it to work very soon using an Arduino Beetle.

Hoping to make a custom JST converter board for use on multiple mixers.

Thanks for the inspiration!